+++ src/hooks/useIsomorphicLayoutEffect.ts
@@ -1,0 +1,7 @@
+import { useEffect, useLayoutEffect as vanillaUseLayoutEffect } from 'react'
+
+export const isBrowser = typeof window !== 'undefined'
+
+export const useLayoutEffect = isBrowser ? vanillaUseLayoutEffect : useEffect
+
+export default useLayoutEffect