+++ src/index.ts
@@ -7,7 +7,6 @@ export { UseCanvas } from './components/UseCanvas'
 // Hooks
 export { useScrollRig } from './hooks/useScrollRig'
 export { useCanvas } from './hooks/useCanvas'
-export { useCanvasRef } from './hooks/useCanvasRef'
 export { useScrollbar } from './scrollbar/useScrollbar'
 export { useTracker } from './hooks/useTracker'
 
@@ -17,7 +16,15 @@ export { useImageAsTexture } from './hooks/useImageAsTexture'
 // Scrollbar
 export { SmoothScrollbar } from './scrollbar/SmoothScrollbar'
 
+// CSS class names for hiding stuff
+// Matching css styles can be imported from @14islands/r3f-scrollr-rig/css
+export const styles = {
+  hidden: 'ScrollRig-visibilityHidden',
+  hiddenWhenSmooth: 'ScrollRig-visibilityHidden ScrollRig-hiddenIfSmooth',
+  transparentColor: 'ScrollRig-transparentColor',
+  transparentColorWhenSmooth: 'ScrollRig-transparentColor ScrollRig-hiddenIfSmooth',
+}
+
 // Private-ish
 // ----------------------------------
 export { useCanvasStore } from './store'
-export { config as _config } from './config'