+++ src/scrollbar/index.js
@@ -1,5 +1,3 @@
 export { useScrollbar } from './useScrollbar'
 export { SmoothScrollbar } from './SmoothScrollbar'
-
 export { useTracker } from '../hooks/useTracker'
-export { useCanvasRef } from '../hooks/useCanvasRef'